Split VSO tests into attrs and client
