+++ sunpy/net/vso/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from sunpy.net import vso
+
+
+@pytest.fixture
+def client():
+    return vso.VSOClient()
